rough first attempt at clientizing the gui
